Updated Data Layer examples

+++ examples/11.data-layer/app/src/main/java/shopapp/view/ShoppingItemScreen.kt
@@ -114,7 +114,7 @@ fun ShoppingItemScreen(onNavigateBack: () -> Unit) {
                 label = "Select a Product",
                 options = products,
                 selectedOptionId = productId,
-                onSelectionChange = { productId = it.toLong() })
+                onSelectionChange = { productId = it })
 
             OutlinedTextField(
                 value = if (quantity > 0) quantity.toString() else "",
